It37g2019 feature

+++ src/Blockcore/NBitcoin/BouncyCastle/crypto/digests/RipeMD160Digest.cs
@@ -258,7 +258,7 @@ namespace NBitcoin.BouncyCastle.Crypto.Digests
             // right
             aa = RL(aa + F5(bb, cc, dd) + this.X[5] + unchecked((int)0x50a28be6), 8) + ee;
             cc = RL(cc, 10);
-            ee = RL(ee + F5(aa, bb, cc) + this.X[14] + unchecked((int)0x50a28be6), 9) + dd;
+            ee = RL(ee + F5(aa, bb, cc) + this.X[14] + unchecked(0x50a28be6), 9) + dd;
             bb = RL(bb, 10);
             dd = RL(dd + F5(ee, aa, bb) + this.X[7] + unchecked((int)0x50a28be6), 9) + cc;
             aa = RL(aa, 10);
